rename fields and move cache
